Struct rusoto_codebuild::Build

pub struct Build {
    pub arn: Option<String>,
    pub artifacts: Option<BuildArtifacts>,
    pub build_complete: Option<bool>,
    pub build_status: Option<String>,
    pub current_phase: Option<String>,
    pub end_time: Option<f64>,
    pub environment: Option<ProjectEnvironment>,
    pub id: Option<String>,
    pub initiator: Option<String>,
    pub logs: Option<LogsLocation>,
    pub phases: Option<Vec<BuildPhase>>,
    pub project_name: Option<String>,
    pub source: Option<ProjectSource>,
    pub source_version: Option<String>,
    pub start_time: Option<f64>,
    pub timeout_in_minutes: Option<i64>,
}

Information about a build.

Fields
arn: Option<String>

The Amazon Resource Name (ARN) of the build.

artifacts: Option<BuildArtifacts>

Information about the output artifacts for the build.

build_complete: Option<bool>

Whether the build has finished. True if completed; otherwise, false.

build_status: Option<String>

The current status of the build. Valid values include:

FAILED: The build failed.

FAULT: The build faulted.

IN_PROGRESS: The build is still in progress.

STOPPED: The build stopped.

SUCCEEDED: The build succeeded.

TIMED_OUT: The build timed out.

current_phase: Option<String>

The current build phase.

end_time: Option<f64>

When the build process ended, expressed in Unix time format.

environment: Option<ProjectEnvironment>

Information about the build environment for this build.

id: Option<String>

The unique ID for the build.

initiator: Option<String>

The entity that started the build. Valid values include:

If AWS CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline).

If an AWS Identity and Access Management (IAM) user started the build, the user's name (for example MyUserName).

If the Jenkins plugin for AWS CodeBuild started the build, the string CodeBuild-Jenkins-Plugin.

logs: Option<LogsLocation>

Information about the build's logs in Amazon CloudWatch Logs.

phases: Option<Vec<BuildPhase>>

Information about all previous build phases that are completed and information about any current build phase that is not yet complete.

project_name: Option<String>

The name of the build project.

source: Option<ProjectSource>

Information about the source code to be built.

source_version: Option<String>

Any version identifier for the version of the source code to be built.

start_time: Option<f64>

When the build process started, expressed in Unix time format.

timeout_in_minutes: Option<i64>

How long, in minutes, for AWS CodeBuild to wait before timing out this build if it does not get marked as completed.