Struct rusoto_codebuild::BuildPhase

pub struct BuildPhase {
    pub contexts: Option<Vec<PhaseContext>>,
    pub duration_in_seconds: Option<i64>,
    pub end_time: Option<f64>,
    pub phase_status: Option<String>,
    pub phase_type: Option<String>,
    pub start_time: Option<f64>,
}

Information about a stage for a build.

Fields
contexts: Option<Vec<PhaseContext>>

Additional information about a build phase, especially to help troubleshoot a failed build.

duration_in_seconds: Option<i64>

How long, in seconds, between the starting and ending times of the build's phase.

end_time: Option<f64>

When the build phase ended, expressed in Unix time format.

phase_status: Option<String>

The current status of the build phase. Valid values include:

FAILED: The build phase failed.

FAULT: The build phase faulted.

IN_PROGRESS: The build phase is still in progress.

STOPPED: The build phase stopped.

SUCCEEDED: The build phase succeeded.

TIMED_OUT: The build phase timed out.

phase_type: Option<String>

The name of the build phase. Valid values include:

BUILD: Core build activities typically occur in this build phase.

COMPLETED: The build has been completed.

DOWNLOAD_SOURCE: Source code is being downloaded in this build phase.

FINALIZING: The build process is completing in this build phase.

INSTALL: Installation activities typically occur in this build phase.

POST_BUILD: Post-build activities typically occur in this build phase.

PRE_BUILD: Pre-build activities typically occur in this build phase.

PROVISIONING: The build environment is being set up.

SUBMITTED: The build has been submitted.

UPLOAD_ARTIFACTS: Build output artifacts are being uploaded to the output location.

start_time: Option<f64>

When the build phase started, expressed in Unix time format.