Struct rusoto_codebuild::CreateProjectInput

pub struct CreateProjectInput {
    pub artifacts: ProjectArtifacts,
    pub description: Option<String>,
    pub encryption_key: Option<String>,
    pub environment: ProjectEnvironment,
    pub name: String,
    pub service_role: Option<String>,
    pub source: ProjectSource,
    pub tags: Option<Vec<Tag>>,
    pub timeout_in_minutes: Option<i64>,
}

Fields
artifacts: ProjectArtifacts

Information about the build output artifacts for the build project.

description: Option<String>

A description that makes the build project easy to identify.

encryption_key: Option<String>

The AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts.

You can specify either the CMK's Amazon Resource Name (ARN) or, if available, the CMK's alias (using the format alias/alias-name ).

environment: ProjectEnvironment

Information about the build environment for the build project.

name: String

The name of the build project.

service_role: Option<String>

The ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

source: ProjectSource

Information about the build input source code for the build project.

tags: Option<Vec<Tag>>

A set of tags for this build project.

These tags are available for use by AWS services that support AWS CodeBuild build project tags.

timeout_in_minutes: Option<i64>

How long, in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait until timing out any build that has not been marked as completed. The default is 60 minutes.