Struct rusoto_codebuild::ProjectArtifacts

pub struct ProjectArtifacts {
    pub encryption_disabled: Option<bool>,
    pub location: Option<String>,
    pub name: Option<String>,
    pub namespace_type: Option<String>,
    pub packaging: Option<String>,
    pub path: Option<String>,
    pub type_: String,
}

Information about the build output artifacts for the build project.

Fields
encryption_disabled: Option<bool>

Set to true if you do not want your output artifacts encrypted. This option is only valid if your artifacts type is Amazon S3. If this is set with another artifacts type, an invalidInputException will be thrown.

location: Option<String>

Information about the build output artifact location, as follows:

If type is set to CODEPIPELINE, then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output locations instead of AWS CodeBuild.

If type is set to NO_ARTIFACTS, then this value will be ignored if specified, because no build output will be produced.

If type is set to S3, this is the name of the output bucket.

name: Option<String>

Along with path and namespaceType, the pattern that AWS CodeBuild will use to name and store the output artifact, as follows:

If type is set to CODEPIPELINE, then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild.

If type is set to NOARTIFACTS, then this value will be ignored if specified, because no build output will be produced.

If type is set to S3, this is the name of the output artifact object. If you set the name to be a forward slash ("/"), then the artifact is stored in the root of the output bucket.

For example:

If path is set to MyArtifacts, namespaceType is set to BUILDID, and name is set to MyArtifact.zip, then the output artifact would be stored in MyArtifacts/build-ID/MyArtifact.zip.

If path is empty, namespaceType is set to NONE, and name is set to "/", then the output artifact would be stored in the root of the output bucket.

If path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to "/", then the output artifact would be stored in MyArtifacts/build-ID .

namespace_type: Option<String>

Along with path and name, the pattern that AWS CodeBuild will use to determine the name and location to store the output artifact, as follows:

If type is set to CODEPIPELINE, then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild.

If type is set to NO_ARTIFACTS, then this value will be ignored if specified, because no build output will be produced.

If type is set to S3, then valid values include:

BUILD_ID: Include the build ID in the location of the build output artifact.

NONE: Do not include the build ID. This is the default if namespaceType is not specified.

For example, if path is set to MyArtifacts, namespaceType is set to BUILD_ID, and name is set to MyArtifact.zip, then the output artifact would be stored in MyArtifacts/build-ID/MyArtifact.zip.

packaging: Option<String>

The type of build output artifact to create, as follows:

If type is set to CODEPIPELINE, then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output artifacts instead of AWS CodeBuild.

If type is set to NO_ARTIFACTS, then this value will be ignored if specified, because no build output will be produced.

If type is set to S3, valid values include:

NONE: AWS CodeBuild will create in the output bucket a folder containing the build output. This is the default if packaging is not specified.

ZIP: AWS CodeBuild will create in the output bucket a ZIP file containing the build output.

path: Option<String>

Along with namespaceType and name, the pattern that AWS CodeBuild will use to name and store the output artifact, as follows:

If type is set to CODEPIPELINE, then AWS CodePipeline will ignore this value if specified. This is because AWS CodePipeline manages its build output names instead of AWS CodeBuild.

If type is set to NO_ARTIFACTS, then this value will be ignored if specified, because no build output will be produced.

If type is set to S3, this is the path to the output artifact. If path is not specified, then path will not be used.

For example, if path is set to MyArtifacts, namespaceType is set to NONE, and name is set to MyArtifact.zip, then the output artifact would be stored in the output bucket at MyArtifacts/MyArtifact.zip.

type_: String

The type of build output artifact. Valid values include:

CODEPIPELINE: The build project will have build output generated through AWS CodePipeline.

NO_ARTIFACTS: The build project will not produce any build output.

S3: The build project will store build output in Amazon Simple Storage Service (Amazon S3).