Struct rusoto_codebuild::ProjectEnvironment

pub struct ProjectEnvironment {
    pub certificate: Option<String>,
    pub compute_type: String,
    pub environment_variables: Option<Vec<EnvironmentVariable>>,
    pub image: String,
    pub privileged_mode: Option<bool>,
    pub type_: String,
}

Information about the build environment of the build project.

Fields
certificate: Option<String>

The certificate to use with this build project.

compute_type: String

Information about the compute resources the build project will use. Available values include:

BUILDGENERAL1SMALL: Use up to 3 GB memory and 2 vCPUs for builds.

BUILDGENERAL1MEDIUM: Use up to 7 GB memory and 4 vCPUs for builds.

BUILDGENERAL1LARGE: Use up to 15 GB memory and 8 vCPUs for builds.

environment_variables: Option<Vec<EnvironmentVariable>>

A set of environment variables to make available to builds for this build project.

image: String

The ID of the Docker image to use for this build project.

privileged_mode: Option<bool>

Enables running the Docker daemon inside a Docker container. Set to true only if the build project is be used to build Docker images, and the specified build environment image is not provided by AWS CodeBuild with Docker support. Otherwise, all associated builds that attempt to interact with the Docker daemon will fail. Note that you must also start the Docker daemon so that builds can interact with it. One way to do this is to initialize the Docker daemon during the install phase of your build spec by running the following build commands. (Do not run the following build commands if the specified build environment image is provided by AWS CodeBuild with Docker support.)

- nohup /usr/local/bin/dockerd --host=unix:///var/run/docker.sock --host=tcp://0.0.0.0:2375 --storage-driver=overlay& - timeout -t 15 sh -c "until docker info; do echo .; sleep 1; done"

type_: String

The type of build environment to use for related builds.