Struct rusoto_codebuild::ProjectSource

pub struct ProjectSource {
    pub auth: Option<SourceAuth>,
    pub buildspec: Option<String>,
    pub location: Option<String>,
    pub type_: String,
}

Information about the build input source code for the build project.

Fields
auth: Option<SourceAuth>

Information about the authorization settings for AWS CodeBuild to access the source code to be built.

This information is for the AWS CodeBuild console's use only. Your code should not get or set this information directly (unless the build project's source type value is GITHUB).

buildspec: Option<String>

The build spec declaration to use for the builds in this build project.

If this value is not specified, a build spec must be included along with the source code to be built.

location: Option<String>

Information about the location of the source code to be built. Valid values include:

For source code settings that are specified in the source action of a pipeline in AWS CodePipeline, location should not be specified. If it is specified, AWS CodePipeline will ignore it. This is because AWS CodePipeline uses the settings in a pipeline's source action instead of this value.

For source code in an AWS CodeCommit repository, the HTTPS clone URL to the repository that contains the source code and the build spec (for example, https://git-codecommit.region-ID.amazonaws.com/v1/repos/repo-name ).

For source code in an Amazon Simple Storage Service (Amazon S3) input bucket, the path to the ZIP file that contains the source code (for example, bucket-name/path/to/object-name.zip)

For source code in a GitHub repository, the HTTPS clone URL to the repository that contains the source and the build spec. Also, you must connect your AWS account to your GitHub account. To do this, use the AWS CodeBuild console to begin creating a build project. When you use the console to connect (or reconnect) with GitHub, on the GitHub Authorize application page that displays, for Organization access, choose Request access next to each repository you want to allow AWS CodeBuild to have access to. Then choose Authorize application. (After you have connected to your GitHub account, you do not need to finish creating the build project, and you may then leave the AWS CodeBuild console.) To instruct AWS CodeBuild to then use this connection, in the source object, set the auth object's type value to OAUTH.

type_: String

The type of repository that contains the source code to be built. Valid values include:

CODECOMMIT: The source code is in an AWS CodeCommit repository.

CODEPIPELINE: The source code settings are specified in the source action of a pipeline in AWS CodePipeline.

GITHUB: The source code is in a GitHub repository.

S3: The source code is in an Amazon Simple Storage Service (Amazon S3) input bucket.