Crate rusoto_codebuild

AWS CodeBuild
AWS CodeBuild is a fully managed build service in the cloud. AWS CodeBuild compiles your source code, runs unit tests, and produces artifacts that are ready to deploy. AWS CodeBuild eliminates the need to provision, manage, and scale your own build servers. It provides prepackaged build environments for the most popular programming languages and build tools, such as Apache Maven, Gradle, and more. You can also fully customize build environments in AWS CodeBuild to use your own build tools. AWS CodeBuild scales automatically to meet peak build requests, and you pay only for the build time you consume. For more information about AWS CodeBuild, see the AWS CodeBuild User Guide.

AWS CodeBuild supports these operations:

BatchDeleteBuilds: Deletes one or more builds.

BatchGetProjects: Gets information about one or more build projects. A build project defines how AWS CodeBuild will run a build. This includes information such as where to get the source code to build, the build environment to use, the build commands to run, and where to store the build output. A build environment represents a combination of operating system, programming language runtime, and tools that AWS CodeBuild will use to run a build. Also, you can add tags to build projects to help manage your resources and costs.

CreateProject: Creates a build project.

CreateWebhook: For an existing AWS CodeBuild build project that has its source code stored in a GitHub repository, enables AWS CodeBuild to begin automatically rebuilding the source code every time a code change is pushed to the repository.

UpdateWebhook: Changes the settings of an existing webhook.

DeleteProject: Deletes a build project.

DeleteWebhook: For an existing AWS CodeBuild build project that has its source code stored in a GitHub repository, stops AWS CodeBuild from automatically rebuilding the source code every time a code change is pushed to the repository.

ListProjects: Gets a list of build project names, with each build project name representing a single build project.

UpdateProject: Changes the settings of an existing build project.

BatchGetBuilds: Gets information about one or more builds.

ListBuilds: Gets a list of build IDs, with each build ID representing a single build.

ListBuildsForProject: Gets a list of build IDs for the specified build project, with each build ID representing a single build.

StartBuild: Starts running a build.

StopBuild: Attempts to stop running a build.

ListCuratedEnvironmentImages: Gets information about Docker images that are managed by AWS CodeBuild.

If you're using the service, you're probably looking for CodeBuildClient and CodeBuild.

Structs

BatchDeleteBuildsInput

BatchDeleteBuildsOutput

BatchGetBuildsInput

BatchGetBuildsOutput

BatchGetProjectsInput

BatchGetProjectsOutput

Build
Information about a build.

BuildArtifacts
Information about build output artifacts.

BuildNotDeleted
Information about a build that could not be successfully deleted.

BuildPhase
Information about a stage for a build.

CodeBuildClient
A client for the AWS CodeBuild API.

CreateProjectInput

CreateProjectOutput

CreateWebhookInput

CreateWebhookOutput

DeleteProjectInput

DeleteProjectOutput

DeleteWebhookInput

DeleteWebhookOutput

EnvironmentImage
Information about a Docker image that is managed by AWS CodeBuild.

EnvironmentLanguage
A set of Docker images that are related by programming language and are managed by AWS CodeBuild.

EnvironmentPlatform
A set of Docker images that are related by platform and are managed by AWS CodeBuild.

EnvironmentVariable
Information about an environment variable for a build project or a build.

InvalidateProjectCacheInput

InvalidateProjectCacheOutput

ListBuildsForProjectInput

ListBuildsForProjectOutput

ListBuildsInput

ListBuildsOutput

ListCuratedEnvironmentImagesInput

ListCuratedEnvironmentImagesOutput

ListProjectsInput

ListProjectsOutput

LogsLocation
Information about build logs in Amazon CloudWatch Logs.

NetworkInterface
Describes a network interface.

PhaseContext
Additional information about a build phase that has an error. You can use this information to help troubleshoot a failed build.

Project
Information about a build project.

ProjectArtifacts
Information about the build output artifacts for the build project.

ProjectBadge
Information about the build badge for the build project.

ProjectCache
Information about the cache for the build project.

ProjectEnvironment
Information about the build environment of the build project.

ProjectSource
Information about the build input source code for the build project.

SourceAuth
Information about the authorization settings for AWS CodeBuild to access the source code to be built.

This information is for the AWS CodeBuild console's use only. Your code should not get or set this information directly (unless the build project's source type value is BITBUCKET or GITHUB).

StartBuildInput

StartBuildOutput

StopBuildInput

StopBuildOutput

Tag
A tag, consisting of a key and a value.

This tag is available for use by AWS services that support tags in AWS CodeBuild.

UpdateProjectInput

UpdateProjectOutput

UpdateWebhookInput

UpdateWebhookOutput

VpcConfig
Information about the VPC configuration that AWS CodeBuild will access.

Webhook
Information about a webhook in GitHub that connects repository events to a build project in AWS CodeBuild.

Enums

BatchDeleteBuildsError
Errors returned by BatchDeleteBuilds

BatchGetBuildsError
Errors returned by BatchGetBuilds

BatchGetProjectsError
Errors returned by BatchGetProjects

CreateProjectError
Errors returned by CreateProject

CreateWebhookError
Errors returned by CreateWebhook

DeleteProjectError
Errors returned by DeleteProject

DeleteWebhookError
Errors returned by DeleteWebhook

InvalidateProjectCacheError
Errors returned by InvalidateProjectCache

ListBuildsError
Errors returned by ListBuilds

ListBuildsForProjectError
Errors returned by ListBuildsForProject

ListCuratedEnvironmentImagesError
Errors returned by ListCuratedEnvironmentImages

ListProjectsError
Errors returned by ListProjects

StartBuildError
Errors returned by StartBuild

StopBuildError
Errors returned by StopBuild

UpdateProjectError
Errors returned by UpdateProject

UpdateWebhookError
Errors returned by UpdateWebhook

Traits

CodeBuild
Trait representing the capabilities of the AWS CodeBuild API. AWS CodeBuild clients implement this trait.