Struct rusoto_codebuild::CodeBuildClient

pub struct CodeBuildClient { /* fields omitted */ }

A client for the AWS CodeBuild API.

Methods

impl CodeBuildClient

pub fn new(region: Region) -> CodeBuildClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CodeBuildClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,