Trait rusoto_codebuild::CodeBuild

pub trait CodeBuild {
    fn batch_delete_builds(
        &self, 
        input: BatchDeleteBuildsInput
    ) -> RusotoFuture<BatchDeleteBuildsOutput, BatchDeleteBuildsError>;

    fn batch_get_builds(
        &self, 
        input: BatchGetBuildsInput
    ) -> RusotoFuture<BatchGetBuildsOutput, BatchGetBuildsError>;

    fn batch_get_projects(
        &self, 
        input: BatchGetProjectsInput
    ) -> RusotoFuture<BatchGetProjectsOutput, BatchGetProjectsError>;

    fn create_project(
        &self, 
        input: CreateProjectInput
    ) -> RusotoFuture<CreateProjectOutput, CreateProjectError>;

    fn create_webhook(
        &self, 
        input: CreateWebhookInput
    ) -> RusotoFuture<CreateWebhookOutput, CreateWebhookError>;

    fn delete_project(
        &self, 
        input: DeleteProjectInput
    ) -> RusotoFuture<DeleteProjectOutput, DeleteProjectError>;

    fn delete_webhook(
        &self, 
        input: DeleteWebhookInput
    ) -> RusotoFuture<DeleteWebhookOutput, DeleteWebhookError>;

    fn invalidate_project_cache(
        &self, 
        input: InvalidateProjectCacheInput
    ) -> RusotoFuture<InvalidateProjectCacheOutput, InvalidateProjectCacheError>;

    fn list_builds(
        &self, 
        input: ListBuildsInput
    ) -> RusotoFuture<ListBuildsOutput, ListBuildsError>;

    fn list_builds_for_project(
        &self, 
        input: ListBuildsForProjectInput
    ) -> RusotoFuture<ListBuildsForProjectOutput, ListBuildsForProjectError>;

    fn list_curated_environment_images(
        &self
    ) -> RusotoFuture<ListCuratedEnvironmentImagesOutput, ListCuratedEnvironmentImagesError>;

    fn list_projects(
        &self, 
        input: ListProjectsInput
    ) -> RusotoFuture<ListProjectsOutput, ListProjectsError>;

    fn start_build(
        &self, 
        input: StartBuildInput
    ) -> RusotoFuture<StartBuildOutput, StartBuildError>;

    fn stop_build(
        &self, 
        input: StopBuildInput
    ) -> RusotoFuture<StopBuildOutput, StopBuildError>;

    fn update_project(
        &self, 
        input: UpdateProjectInput
    ) -> RusotoFuture<UpdateProjectOutput, UpdateProjectError>;

    fn update_webhook(
        &self, 
        input: UpdateWebhookInput
    ) -> RusotoFuture<UpdateWebhookOutput, UpdateWebhookError>;
}

Trait representing the capabilities of the AWS CodeBuild API. AWS CodeBuild clients implement this trait.

Required Methods

fn batch_delete_builds(
&self,
input: BatchDeleteBuildsInput
) -> RusotoFuture<BatchDeleteBuildsOutput, BatchDeleteBuildsError>

Deletes one or more builds.

fn batch_get_builds(
&self,
input: BatchGetBuildsInput
) -> RusotoFuture<BatchGetBuildsOutput, BatchGetBuildsError>

Gets information about builds.

fn batch_get_projects(
&self,
input: BatchGetProjectsInput
) -> RusotoFuture<BatchGetProjectsOutput, BatchGetProjectsError>

Gets information about build projects.

fn create_project(
&self,
input: CreateProjectInput
) -> RusotoFuture<CreateProjectOutput, CreateProjectError>

Creates a build project.

fn create_webhook(
&self,
input: CreateWebhookInput
) -> RusotoFuture<CreateWebhookOutput, CreateWebhookError>

For an existing AWS CodeBuild build project that has its source code stored in a GitHub repository, enables AWS CodeBuild to begin automatically rebuilding the source code every time a code change is pushed to the repository.

If you enable webhooks for an AWS CodeBuild project, and the project is used as a build step in AWS CodePipeline, then two identical builds will be created for each commit. One build is triggered through webhooks, and one through AWS CodePipeline. Because billing is on a per-build basis, you will be billed for both builds. Therefore, if you are using AWS CodePipeline, we recommend that you disable webhooks in CodeBuild. In the AWS CodeBuild console, clear the Webhook box. For more information, see step 5 in Change a Build Project's Settings.

fn delete_project(
&self,
input: DeleteProjectInput
) -> RusotoFuture<DeleteProjectOutput, DeleteProjectError>

Deletes a build project.

fn delete_webhook(
&self,
input: DeleteWebhookInput
) -> RusotoFuture<DeleteWebhookOutput, DeleteWebhookError>

For an existing AWS CodeBuild build project that has its source code stored in a GitHub repository, stops AWS CodeBuild from automatically rebuilding the source code every time a code change is pushed to the repository.

fn invalidate_project_cache(
&self,
input: InvalidateProjectCacheInput
) -> RusotoFuture<InvalidateProjectCacheOutput, InvalidateProjectCacheError>

Resets the cache for a project.

fn list_builds(
&self,
input: ListBuildsInput
) -> RusotoFuture<ListBuildsOutput, ListBuildsError>

Gets a list of build IDs, with each build ID representing a single build.

fn list_builds_for_project(
&self,
input: ListBuildsForProjectInput
) -> RusotoFuture<ListBuildsForProjectOutput, ListBuildsForProjectError>

Gets a list of build IDs for the specified build project, with each build ID representing a single build.

fn list_curated_environment_images(
&self
) -> RusotoFuture<ListCuratedEnvironmentImagesOutput, ListCuratedEnvironmentImagesError>

Gets information about Docker images that are managed by AWS CodeBuild.

fn list_projects(
&self,
input: ListProjectsInput
) -> RusotoFuture<ListProjectsOutput, ListProjectsError>

Gets a list of build project names, with each build project name representing a single build project.

fn start_build(
&self,
input: StartBuildInput
) -> RusotoFuture<StartBuildOutput, StartBuildError>

Starts running a build.

fn stop_build(
&self,
input: StopBuildInput
) -> RusotoFuture<StopBuildOutput, StopBuildError>

Attempts to stop running a build.

fn update_project(
&self,
input: UpdateProjectInput
) -> RusotoFuture<UpdateProjectOutput, UpdateProjectError>

Changes the settings of a build project.

fn update_webhook(
&self,
input: UpdateWebhookInput
) -> RusotoFuture<UpdateWebhookOutput, UpdateWebhookError>

Updates the webhook associated with an AWS CodeBuild build project.